requirements.txt
streamlit==1.31.1
torch==2.7.1
torchvision==0.22.1
opencv-python==4.7.0.72
pillow==9.5.0
numpy==1.24.3
requests==2.28.2
scikit-image==0.22.0
kaggle==1.5.16 